QA-1363:TechDebt - rename txma test script file
